chore: Add configuration and make target for linters
This adds the configuration to install and run golangci-lint to manage
the execution of multiple linters.

To install the linters run:

    make golangci-lint

Or to install it along with other tools:

    make tools

Then to run the linters use:

    make lint

To run the linters just against new changes, run

    make lint-new

This can be useful to check new changes in a PR for lint issues.

The linters can be configured to run in an editor as well.

See:
    https://golangci-lint.run/usage/integrations/#editor-integration